+++ internal/interfaces/controllers/tweetController/tweet_controller.go
@@ -1,21 +1,20 @@
 package tweetController
 
 import (
-	"context"
 	"net/http"
 	"strconv"
-	"time"
 	"uala-challenge/internal/domain/tweets"
+	"uala-challenge/internal/services/tweetsService"
 
 	"github.com/gin-gonic/gin"
 )
 
 type TweetController struct {
-	tweetRepo tweets.TweetRepository
+	tweetService tweetsService.TweetsService
 }
 
-func NewTweetController(tweetRepo tweets.TweetRepository) *TweetController {
-	return &TweetController{tweetRepo: tweetRepo}
+func NewTweetController(tweetService tweetsService.TweetsService) *TweetController {
+	return &TweetController{tweetService: tweetService}
 }
 
 func (tc *TweetController) CreateTweet(c *gin.Context) {
@@ -25,28 +24,22 @@ func (tc *TweetController) CreateTweet(c *gin.Context) {
 		Content string `json:"content" binding:"required"`
 	}
 
+	// parameter validation and parsing
 	if err := c.ShouldBindJSON(&tweetRequest); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	var tweet tweets.Tweet
-
-	tweet.UserId = tweetRequest.UserId
-	tweet.Content = tweetRequest.Content
-	tweet.CreationDate = time.Now()
-
-	if len(tweet.Content) > tweets.MAX_TWEET_LENGTH {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Tweet content is too long"})
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	tweet, err := tc.tweetRepo.CreateTweet(ctx, &tweet)
+	tweet, err := tc.tweetService.CreateTweet(tweetRequest.UserId, tweetRequest.Content)
 
 	if err != nil {
+
+		// business error handling
+		if err == tweets.ErrTweetTooLong {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
@@ -55,8 +48,10 @@ func (tc *TweetController) CreateTweet(c *gin.Context) {
 }
 
 func (tc *TweetController) GetTweets(c *gin.Context) {
+
 	userId_param := c.Param("userId")
 
+	// parameter parse and validation
 	userId, err := strconv.Atoi(userId_param)
 
 	if err != nil {
@@ -64,20 +59,19 @@ func (tc *TweetController) GetTweets(c *gin.Context) {
 		return
 	}
 
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	tweets, err := tc.tweetRepo.GetTweetsByUserId(ctx, userId)
+	tweetList, err := tc.tweetService.GetTweetsByUserId(userId)
 
 	if err != nil {
+
+		// business error handling
+		if err == tweets.ErrEmptyTimeline {
+			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
+			return
+		}
+
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
 
-	if len(tweets) == 0 {
-		c.JSON(http.StatusNotFound, gin.H{"error": "No tweets found"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"payload": tweets})
+	c.JSON(http.StatusOK, gin.H{"payload": tweetList})
 }